feat(fro-24): complaints form page completed
