--- frontend/client/src/Pages/FileComplaintsOld/components/card/data.js
@@ -1,45 +1,0 @@
-import bulb from "../assets/light.png";
-import calculator from "../assets/calculator.png";
-import question from "../assets/question.png";
-import thumb from "../assets/thumb.png";
-import Vector from "../assets/Vector.png";
-import i from "../assets/i.png";
-
-export const data = [
-  {
-    id: 1,
-    img: bulb,
-    title: "Knowledge base",
-    desc: "It is a long established fact that areader will be distracted by thereadable content of a page when.",
-  },
-  {
-    id: 2,
-    img: Vector,
-    title: "Billing & licensing",
-    desc: "It is a long established fact that areader will be distracted by thereadable content of a page when.",
-  },
-  {
-    id: 3,
-    img: i,
-    title: "General Help",
-    desc: "It is a long established fact that areader will be distracted by thereadable content of a page when.",
-  },
-  {
-    id: 4,
-    img: calculator,
-    title: "Account",
-    desc: "It is a long established fact that areader will be distracted by thereadable content of a page when.",
-  },
-  {
-    id: 5,
-    img: thumb,
-    title: "Reviews",
-    desc: "It is a long established fact that areader will be distracted by thereadable content of a page when.",
-  },
-  {
-    id: 6,
-    img: question,
-    title: "Faqs",
-    desc: "It is a long established fact that areader will be distracted by thereadable content of a page when.",
-  },
-];
