--- frontend/client/src/Pages/FileComplaintsOld/components/complaint-form/ComplaintForm.styled.js
@@ -1,18 +1,0 @@
-import styled from "styled-components";
-
-export const StyledComplaintForm = styled.section`
-  width: 100%;
-  padding: 50px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  & h3 {
-    font-size: 40px;
-    font-weight: bold;
-  }
-
-  & p {
-    font-size: 18px;
-  }
-`;
